renamed CaseService to CaseClient
